+++ src/main/java/com/linker/product/dao/ProductDAO.java
@@ -18,6 +18,6 @@ public interface ProductDAO {
 	public List<ProductDTO> selectProduct(int product_no) throws DataAccessException;
 	
 	// user_id에 해당하는 상품정보 추출하기
-	public List<ProductDTO> read(int user_id) throws DataAccessException;
+	public List<ProductDTO> read(Long user_id) throws DataAccessException;
 
 }
