+++ src/main/java/com/linker/review/dto/ReviewDTO.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class ReviewDTO {
 	
 	private int review_no;	// 리뷰 고유 번호
-	private int user_id;	// 유저 고유 번호
+	private Long user_id;	// 유저 고유 번호
 	private int product_no;	// 리뷰할 상품 번호
 	private int rating;	// 구매 평점
 	private String img_url;	// 상품 이미지url
