+++ src/main/java/com/linker/util/address/AddressService.java
@@ -13,7 +13,7 @@ public class AddressService {
 	private AddressDAO addressDAO;
 	
 	// 아이디에 해당하는 주소지 리스트
-	public List<AddressDTO> addressList(int user_id) throws Exception {
+	public List<AddressDTO> addressList(Long user_id) throws Exception {
 		List<AddressDTO> addressDTO = addressDAO.addressList(user_id);
 		return addressDTO;
 	}
